hacking wallet service endpoint into grant
