+++ cmd/rewards.go
@@ -1,19 +1,10 @@
 package cmd
 
 import (
-	"context"
-	"time"
 
 	// pprof imports
 	_ "net/http/pprof"
 
-	"github.com/brave-intl/bat-go/middleware"
-	appctx "github.com/brave-intl/bat-go/utils/context"
-	"github.com/brave-intl/bat-go/utils/handlers"
-	"github.com/brave-intl/bat-go/utils/logging"
-	"github.com/go-chi/chi"
-	chiware "github.com/go-chi/chi/middleware"
-	"github.com/rs/zerolog/hlog"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -58,46 +49,6 @@ func init() {
 		"the default ac choices for the rewards system")
 	must(viper.BindPFlag("default-ac-choices", rewardsCmd.PersistentFlags().Lookup("default-ac-choices")))
 	must(viper.BindEnv("default-ac-choices", "DEFAULT_AC_CHOICES"))
-}
-
-func setupRouter(ctx context.Context) *chi.Mux {
-	logger, err := appctx.GetLogger(ctx)
-	if err != nil {
-		// no logger on context, make a new one
-		ctx, logger = logging.SetupLogger(ctx)
-	}
-
-	r := chi.NewRouter()
-	r.Use(
-		chiware.RequestID,
-		chiware.RealIP,
-		chiware.Heartbeat("/"),
-		chiware.Timeout(10*time.Second),
-		middleware.BearerToken,
-		middleware.RateLimiter(ctx),
-		middleware.RequestIDTransfer)
-	if logger != nil {
-		// Also handles panic recovery
-		r.Use(
-			hlog.NewHandler(*logger),
-			hlog.UserAgentHandler("user_agent"),
-			hlog.RequestIDHandler("req_id", "Request-Id"),
-			middleware.RequestLogger(logger))
-
-		logger.Info().
-			Str("version", version).
-			Str("commit", commit).
-			Str("build_time", buildTime).
-			Str("ratios_service", viper.GetString("ratios-service")).
-			Str("ledger_service", viper.GetString("ledger-service")).
-			Str("address", viper.GetString("address")).
-			Str("environment", viper.GetString("environment")).
-			Msg("server starting")
-	}
-	// we will always have metrics and health-check
-	r.Get("/metrics", middleware.Metrics())
-	r.Get("/health-check", handlers.HealthCheckHandler(version, buildTime, commit))
-	return r
 }
 
 var (
